cv-service/pyproject.toml
[project]
name = "cv-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "fastapi>=0.135.1",
    "mypy>=1.19.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.6",
    "uvicorn>=0.42.0",
]
